lab1

+++ lab1/lab1.h
@@ -26,7 +26,7 @@ class Lab1VideoGenerator {
 	float * loose_noise;
 	float * dense_noise;
 	vector<Particle> particles;
-	int freq = 8;
+	float freq = 8;
 public:
 	Lab1VideoGenerator();
 	~Lab1VideoGenerator();
